Now that annotations are rendered into the DOM, we can animate them in the `"frameChange"` event handler, which would first attempt to get X and Y positions for each node. #36
Visibility on stage is driven by `start` and `end` members defining a range of frames in which the annotation is visible. For example `start:5, end:7` keep the annotation visible for frames 5, 6 and 7 and hidden for the rest of frames.

Position on stage is defined by `x` and `y` annotation object members. It can be defined as a single numeric value and in that case the annotation will hold the same position. Or it can be an array of numeric values for `start` to `end`, whose length is equal to the visibility range length. For example having the same `start` and `end` as above, setting `x:[ 20, 40, 60 ]` would place the annotation on 20 for frame 5, 40 for frame 6 and 60 for 7. Both X and Y coordinates need to be resolved in order to have the annotation visible.